Revert "HBX-1739: Move class 'org.hibernate.tool.internal.export.pojo.POJOExporter' to 'org.hibernate.tool.api.export.PojoExporter'"
This reverts commit 1e6fabbf1b605ddf13dc8d769d8db323cdb8cd4b.
