+++ orm/src/main/java/org/hibernate/tool/internal/export/dao/DAOExporter.java
@@ -2,10 +2,10 @@ package org.hibernate.tool.internal.export.dao;
 
 import java.util.Map;
 
-import org.hibernate.tool.api.export.PojoExporter;
 import org.hibernate.tool.internal.export.pojo.POJOClass;
+import org.hibernate.tool.internal.export.pojo.POJOExporter;
 
-public class DAOExporter extends PojoExporter {
+public class DAOExporter extends POJOExporter {
 
     private static final String DAO_DAOHOME_FTL = "dao/daohome.ftl";
