+++ test/common/src/main/resources/org/hibernate/tool/ant/GenericExport/build.xml
@@ -32,7 +32,7 @@
 				filepattern="{package-name}/{class-name}.quote" />
 			
 			<hbmtemplate 
-				exporterclass="org.hibernate.tool.api.export.PojoExporter" 
+				exporterclass="org.hibernate.tool.internal.export.pojo.POJOExporter" 
 				filepattern="{package-name}/{class-name}.pojo" /> 
 			
 		</hibernatetool>
